backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agno-admet-backend"
version = "0.1.0"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-multipart>=0.0.9",
  "rdkit-pypi>=2022.9.5",
  "agno>=1.7",
  "admet-ai>=0.2.6; python_version>='3.9'"
] 